backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
requests==2.31.0
pdfplumber==0.9.0
pytesseract==0.3.10
Pillow==9.5.0
qdrant-client==1.0.0
numpy==1.25.0
sentence-transformers==2.2.2
transformers==4.31.0
torch==2.0.1
openai==0.27.0
python-dotenv==1.0.0
